move frontend files into appropriate folders; update imports
