+++ frontend/src/types/typesAndInterfaces.ts
